getting rid of some api branch cruft
